backend/requirements.txt
alembic
fastapi
passlib[bcrypt]
psycopg2
pydantic
pyjwt
python-decouple
python-multipart
sqlmodel
uvicorn